fix(cli): unify focused child status copy
