+++ packages/cli/src/chat/tui/panes/SubagentList.tsx
@@ -7,7 +7,6 @@ import { useMemo } from 'react';
 // Local imports - shared stream state
 import {
   AgentCategory,
-  WORKFLOW_TASK_STATUS_LABEL,
   type StreamTabId,
   type WorkflowCallProgress,
 } from '@shared/schemas';
@@ -19,13 +18,12 @@ import {
 import {
   formatPhaseStageLabel,
   formatRoundStageLabel,
-  formatStreamStatusLabel,
-  streamStatusDisplayKey,
 } from '@shared/streams/streamStatusDisplay';
 import { formatResultCount } from '@utils/text/stringUtils';
 
 // Local imports - TUI rendering
 import { truncateSummaryToWidth } from '../render/terminalText';
+import { formatCliStatusLabel } from '../sessionStatus';
 
 // Local imports - TUI state and controls
 import { childElapsed } from '../state/childControls';
@@ -135,22 +133,11 @@ function SessionRow({
 }): React.JSX.Element {
   const status = session.slice?.status;
   const substate = session.slice?.substate;
-  const statusKey = streamStatusDisplayKey(status, substate);
-  const childStatusLabel =
-    session.parentId !== undefined &&
-    statusKey !== undefined &&
-    Object.hasOwn(WORKFLOW_TASK_STATUS_LABEL, statusKey)
-      ? WORKFLOW_TASK_STATUS_LABEL[
-          statusKey as keyof typeof WORKFLOW_TASK_STATUS_LABEL
-        ]
-      : undefined;
-  const statusLabel =
-    childStatusLabel ??
-    formatStreamStatusLabel(status, {
-      style: 'cli',
-      isChildStream: session.parentId !== undefined,
-      ...(substate ? { substate } : {}),
-    });
+  const statusLabel = formatCliStatusLabel(
+    status,
+    substate,
+    session.parentId !== undefined,
+  );
   const elapsed = childElapsed(
     {
       status,
